Trait integrators::GeneralIntegrator

pub trait GeneralIntegrator<F: Float> {
    fn update(&self, t: F, x: &[F]) -> Result<Vec<F>>;

    fn integrate(
        &self, 
        x: &[F], 
        t_0: F, 
        h: F, 
        n: usize, 
        integrator: Integrators
    ) -> Result<Result<F>> { ... }

    fn prepare(&self, x: Vec<F>, n: usize, _: &Token) -> Result<F> { ... }

    fn runge_kutta_4th(
        &self, 
        t_0: F, 
        h: F, 
        n: usize, 
        result: &mut Result<F>, 
        _: &Token
    ) -> Result<()> { ... }
}

A general integrator for a system of 1st-order ODEs

Required methods

fn update(&self, t: F, x: &[F]) -> Result<Vec<F>>

Update the current state as defined by a system of 1st-order ODEs, return the result

Arguments:

t — Current time moment;

x — Current state of the system.

Provided methods

fn integrate(
&self,
x: &[F],
t_0: F,
h: F,
n: usize,
integrator: Integrators
) -> Result<Result<F>>

Integrate the system of 1st-order ODEs

Arguments:

x — Vector of initial values;

t_0 — Initial value of time;

h — Time step;

n — Number of iterations;

integrator — Integration method.

fn prepare(&self, x: Vec<F>, n: usize, _: &Token) -> Result<F>

Prepare a matrix for the result and put the initial values in the first column

Arguments:

x — Vector of initial values;

n — Number of iterations;

token — Private token.

fn runge_kutta_4th(
&self,
t_0: F,
h: F,
n: usize,
result: &mut Result<F>,
_: &Token
) -> Result<()>

Integrate the system using the 4th-order Runge-Kutta method

Arguments:

t_0 — Initial value of time;

h — Time step;

n — Number of iterations;

result — Result matrix;

token — Private token.